Color Selection for Shapes
Added in color selection tool for shapes as well, supporting up to 27 colors at a time.
